+++ src/test/java/persistence/sql/ddl/ColumnTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import persistence.domain.Person;
+import persistence.sql.domain.Column;
+import persistence.sql.domain.DataType;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
